Denormalize Operators utility class
